ENH: Move standaloneRoutine param notebook to its own class
